// src/app/(protected)/organizer/create-event/[id]/libs/interface/paymentForm.interface.ts
export interface PaymentForm {
    id: string;
    accName: string;
    accNum: string;
    bankName: string;
    bankBranch: string;
    typeBusiness: string;
    perName: string;
    perAddress: string;
    taxCode: string;
    companyName: string;
    companyAddress: string;
    companyTaxCode: string;
}